v2/etl_simple/cdk/components/ggAccel.etl_simple.load/artifacts/ggAccel.etl_simple.load/1.0.0/requirements.txt
awsiotsdk
numpy